@@ -3,7 +3,7 @@
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
-    "start": "ng serve",
+    "start": "ng serve --liveReload=false",
     "build": "ng build",
     "test": "ng test",
     "lint": "ng lint",
@@ -16,51 +16,55 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^6.1.10",
-    "@angular/cdk": "^7.2.1",
-    "@angular/common": "^6.1.10",
-    "@angular/compiler": "^6.1.10",
-    "@angular/core": "^6.1.10",
-    "@angular/forms": "^6.1.10",
-    "@angular/http": "^6.1.10",
-    "@angular/material": "^7.2.1",
-    "@angular/platform-browser": "^6.1.10",
-    "@angular/platform-browser-dynamic": "^6.1.10",
-    "@angular/router": "^6.1.10",
-    "@ng-bootstrap/ng-bootstrap": "^3.3.1",
+    "@angular/animations": "^7.2.2",
+    "@angular/cdk": "^7.2.2",
+    "@angular/common": "^7.2.2",
+    "@angular/compiler": "^7.2.2",
+    "@angular/core": "^7.2.2",
+    "@angular/forms": "^7.2.2",
+    "@angular/http": "^7.2.2",
+    "@angular/material": "^7.2.2",
+    "@angular/platform-browser": "^7.2.2",
+    "@angular/platform-browser-dynamic": "^7.2.2",
+    "@angular/router": "^7.2.2",
+    "@ng-bootstrap/ng-bootstrap": "^4.0.1",
     "@types/jquery": "^3.3.29",
     "core-js": "^2.6.1",
     "cors": "^2.8.5",
     "express": "^4.16.4",
     "formidable": "^1.2.1",
+    "mobx": "^5.7.0",
+    "mobx-angular": "^3.0.1",
     "multer": "^1.4.1",
+    "ngx-mobx": "^1.2.2",
     "node-json-db": "^0.9.2",
     "protractor-screenshoter-plugin": "^0.10.3",
-    "rxjs": "~6.2.0",
-    "zone.js": "~0.8.26"
+    "rxjs": "~6.3.3",
+    "tslib": "^1.9.0",
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.8.9",
-    "@angular/cli": "^6.2.9",
-    "@angular/compiler-cli": "^6.1.10",
-    "@angular/language-service": "^6.1.10",
-    "@types/jasmine": "^2.8.14",
+    "@angular-devkit/build-angular": "~0.12.0",
+    "@angular/cli": "^7.2.2",
+    "@angular/compiler-cli": "^7.2.2",
+    "@angular/language-service": "^7.2.2",
+    "@types/jasmine": "^3.3.7",
     "@types/jasminewd2": "^2.0.6",
-    "@types/node": "~8.9.4",
-    "codelyzer": "~4.3.0",
+    "@types/node": "~10.12.18",
+    "codelyzer": "~4.5.0",
     "concurrently": "^4.1.0",
     "cypress": "^3.1.4",
-    "jasmine-core": "~2.99.1",
+    "jasmine-core": "~3.3.0",
     "jasmine-spec-reporter": "~4.2.1",
     "json-server": "^0.14.2",
-    "karma": "~3.0.0",
+    "karma": "~3.1.4",
     "karma-chrome-launcher": "~2.2.0",
     "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.0",
     "protractor": "^5.4.2",
-    "ts-node": "~7.0.0",
-    "tslint": "~5.11.0",
-    "typescript": "~2.9.2"
+    "ts-node": "~8.0.1",
+    "tslint": "~5.12.1",
+    "typescript": "^3.2.4"
   }
 }
