@@ -1,1 +1,1 @@
-angular.module('reporter').constant('data',{"tests":[{"start":"2019-01-17T12:41:36.432Z","specScreenshots":[],"specLogs":[],"specHtmls":[],"failedExpectations":[],"passedExpectations":[{"matcherName":"toEqual","passed":true,"message":"Passed.","actual":"Одежда, обувь, аксессуары","expected":"Одежда, обувь, аксессуары","error":{},"screenshots":[{"img":"screenshots/3c0391c0-1a55-11e9-8b91-5b3fc40b3870.png","browser":"chrome","when":"2019-01-17T12:41:39.570Z"}],"logs":[],"htmls":[],"when":"2019-01-17T12:41:39.291Z","stack":""}],"end":"2019-01-17T12:41:39.293Z","diff":2861,"timeout":30000,"id":"spec0","description":"should display text on first tree node","fullName":"myHomeStore should display text on first tree node","pendingReason":"","status":"passed"},{"start":"2019-01-17T12:41:39.297Z","specScreenshots":[],"specLogs":[],"specHtmls":[],"failedExpectations":[],"passedExpectations":[{"matcherName":"toBeTruthy","passed":true,"message":"Passed.","actual":true,"expected":[],"error":{},"screenshots":[{"img":"screenshots/3def8340-1a55-11e9-8b91-5b3fc40b3870.png","browser":"chrome","when":"2019-01-17T12:41:42.790Z"}],"logs":[],"htmls":[],"when":"2019-01-17T12:41:42.516Z","stack":""},{"matcherName":"toEqual","passed":true,"message":"Passed.","actual":"Coat Blue","expected":"Coat Blue","error":{},"screenshots":[{"img":"screenshots/3e496310-1a55-11e9-8b91-5b3fc40b3870.png","browser":"chrome","when":"2019-01-17T12:41:43.375Z"}],"logs":[],"htmls":[],"when":"2019-01-17T12:41:43.105Z","stack":""}],"end":"2019-01-17T12:41:43.106Z","diff":3809,"timeout":30000,"id":"spec1","description":"should open category tree and see loaded good in the table","fullName":"myHomeStore should open category tree and see loaded good in the table","pendingReason":"","status":"passed"},{"start":"2019-01-17T12:41:43.107Z","specScreenshots":[],"specLogs":[],"specHtmls":[],"failedExpectations":[],"passedExpectations":[{"matcherName":"toBeFalsy","passed":true,"message":"Passed.","actual":false,"expected":[],"error":{},"screenshots":[{"img":"screenshots/40745960-1a55-11e9-8b91-5b3fc40b3870.png","browser":"chrome","when":"2019-01-17T12:41:47.006Z"}],"logs":[],"htmls":[],"when":"2019-01-17T12:41:46.742Z","stack":""}],"end":"2019-01-17T12:41:46.743Z","diff":3636,"timeout":30000,"id":"spec2","description":"should not see good in the table after change active category in tree","fullName":"myHomeStore should not see good in the table after change active category in tree","pendingReason":"","status":"passed"},{"start":"2019-01-17T12:41:46.743Z","specScreenshots":[],"specLogs":[],"specHtmls":[],"failedExpectations":[],"passedExpectations":[{"matcherName":"toBeTruthy","passed":true,"message":"Passed.","actual":true,"expected":[],"error":{},"screenshots":[{"img":"screenshots/4227aeb0-1a55-11e9-8b91-5b3fc40b3870.png","browser":"chrome","when":"2019-01-17T12:41:49.862Z"}],"logs":[],"htmls":[],"when":"2019-01-17T12:41:49.594Z","stack":""},{"matcherName":"toBeTruthy","passed":true,"message":"Passed.","actual":true,"expected":[],"error":{},"screenshots":[{"img":"screenshots/42ace440-1a55-11e9-8b91-5b3fc40b3870.png","browser":"chrome","when":"2019-01-17T12:41:50.763Z"}],"logs":[],"htmls":[],"when":"2019-01-17T12:41:50.468Z","stack":""},{"matcherName":"toEqual","passed":true,"message":"Passed.","actual":"Coat transparent","expected":"Coat transparent","error":{},"screenshots":[{"img":"screenshots/42fd0010-1a55-11e9-8b91-5b3fc40b3870.png","browser":"chrome","when":"2019-01-17T12:41:51.311Z"}],"logs":[],"htmls":[],"when":"2019-01-17T12:41:50.993Z","stack":""}],"end":"2019-01-17T12:41:50.994Z","diff":4251,"timeout":30000,"id":"spec3","description":"should see good in the table after change page","fullName":"myHomeStore should see good in the table after change page","pendingReason":"","status":"passed"},{"start":"2019-01-17T12:41:50.995Z","specScreenshots":[],"specLogs":[],"specHtmls":[],"failedExpectations":[],"passedExpectations":[{"matcherName":"toBeTruthy","passed":true,"message":"Passed.","actual":true,"expected":[],"error":{},"screenshots":[{"img":"screenshots/44bab5a0-1a55-11e9-8b91-5b3fc40b3870.png","browser":"chrome","when":"2019-01-17T12:41:54.201Z"}],"logs":[],"htmls":[],"when":"2019-01-17T12:41:53.914Z","stack":""},{"matcherName":"toEqual","passed":true,"message":"Passed.","actual":"Coat Blue","expected":"Coat Blue","error":{},"screenshots":[{"img":"screenshots/450aaa60-1a55-11e9-8b91-5b3fc40b3870.png","browser":"chrome","when":"2019-01-17T12:41:54.707Z"}],"logs":[],"htmls":[],"when":"2019-01-17T12:41:54.437Z","stack":""},{"matcherName":"toEqual","passed":true,"message":"Passed.","actual":"\"http://localhost:4200/assets/images/coat1.jpg\"","expected":"\"http://localhost:4200/assets/images/coat1.jpg\"","error":{},"screenshots":[{"img":"screenshots/45406060-1a55-11e9-8b91-5b3fc40b3870.png","browser":"chrome","when":"2019-01-17T12:41:55.086Z"}],"logs":[],"htmls":[],"when":"2019-01-17T12:41:54.790Z","stack":""}],"end":"2019-01-17T12:41:54.790Z","diff":3795,"timeout":30000,"id":"spec4","description":"should see in details image equal to clicked in table data","fullName":"myHomeStore should see in details image equal to clicked in table data","pendingReason":"","status":"passed"},{"start":"2019-01-17T12:41:54.791Z","specScreenshots":[],"specLogs":[],"specHtmls":[],"failedExpectations":[],"passedExpectations":[{"matcherName":"toBeTruthy","passed":true,"message":"Passed.","actual":true,"expected":[],"error":{},"screenshots":[{"img":"screenshots/470e90b0-1a55-11e9-8b91-5b3fc40b3870.png","browser":"chrome","when":"2019-01-17T12:41:58.121Z"}],"logs":[],"htmls":[],"when":"2019-01-17T12:41:57.818Z","stack":""},{"matcherName":"toEqual","passed":true,"message":"Passed.","actual":"Coat Blue","expected":"Coat Blue","error":{},"screenshots":[{"img":"screenshots/473fd9e0-1a55-11e9-8b91-5b3fc40b3870.png","browser":"chrome","when":"2019-01-17T12:41:58.406Z"}],"logs":[],"htmls":[],"when":"2019-01-17T12:41:58.142Z","stack":""},{"matcherName":"toEqual","passed":true,"message":"Passed.","actual":"\"Coat Blue\"","expected":"\"Coat Blue\"","error":{},"screenshots":[{"img":"screenshots/47a29350-1a55-11e9-8b91-5b3fc40b3870.png","browser":"chrome","when":"2019-01-17T12:41:59.083Z"}],"logs":[],"htmls":[],"when":"2019-01-17T12:41:58.789Z","stack":""}],"end":"2019-01-17T12:41:58.790Z","diff":3999,"timeout":30000,"id":"spec5","description":"should see in details name of good equal to clicked in table data","fullName":"myHomeStore should see in details name of good equal to clicked in table data","pendingReason":"","status":"passed"}],"stat":{"passed":6,"failed":0,"pending":0,"disabled":0},"ci":{},"generatedOn":"2019-01-17T12:41:59.804Z"});
+angular.module('reporter').constant('data',{"tests":[{"start":"2019-01-22T11:41:43.720Z","specScreenshots":[],"specLogs":[],"specHtmls":[],"failedExpectations":[],"passedExpectations":[],"end":"2019-01-22T11:41:43.722Z","diff":2,"timeout":30000,"id":"spec0","description":"should display text on first tree node","fullName":"myHomeStore should display text on first tree node","pendingReason":"Temporarily disabled with xit","status":"pending"},{"start":"2019-01-22T11:41:43.725Z","specScreenshots":[],"specLogs":[],"specHtmls":[],"failedExpectations":[],"passedExpectations":[],"end":"2019-01-22T11:41:43.726Z","diff":1,"timeout":30000,"id":"spec1","description":"should open category tree and see loaded good in the table","fullName":"myHomeStore should open category tree and see loaded good in the table","pendingReason":"Temporarily disabled with xit","status":"pending"},{"start":"2019-01-22T11:41:43.726Z","specScreenshots":[],"specLogs":[],"specHtmls":[],"failedExpectations":[],"passedExpectations":[],"end":"2019-01-22T11:41:43.726Z","diff":0,"timeout":30000,"id":"spec2","description":"should not see good in the table after change active category in tree","fullName":"myHomeStore should not see good in the table after change active category in tree","pendingReason":"Temporarily disabled with xit","status":"pending"},{"start":"2019-01-22T11:41:43.727Z","specScreenshots":[],"specLogs":[],"specHtmls":[],"failedExpectations":[],"passedExpectations":[],"end":"2019-01-22T11:41:43.728Z","diff":1,"timeout":30000,"id":"spec3","description":"should see good in the table after change page","fullName":"myHomeStore should see good in the table after change page","pendingReason":"Temporarily disabled with xit","status":"pending"},{"start":"2019-01-22T11:41:43.729Z","specScreenshots":[],"specLogs":[],"specHtmls":[],"failedExpectations":[{"matcherName":"","passed":false,"expected":"","actual":"","error":{},"screenshots":[],"logs":[],"htmls":[],"when":"2019-01-22T11:43:32.418Z","message":"Error: Timeout - Async callback was not invoked within timeout specified by jasmine.DEFAULT_TIMEOUT_INTERVAL.","stack":"Error: Timeout - Async callback was not invoked within timeout specified by jasmine.DEFAULT_TIMEOUT_INTERVAL.\n    at Timeout._onTimeout (C:\\Users\\Andro\\Documents\\sid\\angular\\MyHomeStore\\node_modules\\jasmine\\node_modules\\jasmine-core\\lib\\jasmine-core\\jasmine.js:4281:23)\n    at ontimeout (timers.js:436:11)\n    at tryOnTimeout (timers.js:300:5)\n    at listOnTimeout (timers.js:263:5)\n    at Timer.processTimers (timers.js:223:10)"}],"passedExpectations":[],"end":"2019-01-22T11:43:32.421Z","diff":108692,"timeout":30000,"id":"spec4","description":"should see in details image equal to clicked in table data","fullName":"myHomeStore should see in details image equal to clicked in table data","pendingReason":"","status":"failed"}],"stat":{"passed":0,"failed":1,"pending":4,"disabled":0},"ci":{},"generatedOn":"2019-01-22T11:43:32.472Z"});
