@@ -2,14 +2,14 @@ import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 
 import { AppComponent } from './app.component';
-import { CatalogComponent } from './catalog/catalog.component';
-import { GoodsComponent } from './goods/goods.component';
-import { DetailsComponent } from './details/details.component';
+import { CatalogComponent } from './components/catalog/catalog.component';
+import { GoodsComponent } from './components/goods/goods.component';
+import { DetailsComponent } from './components/details/details.component';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { TreeComponent } from './catalog/tree/tree.component';
+import { TreeComponent } from './components/catalog/tree/tree.component';
 
 import { HttpClientModule } from '@angular/common/http';
-import { SimulateDbService } from './simulate-db.service';
+import { DbService } from './services/db.service';
 
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { MatProgressSpinnerModule } from "@angular/material";
@@ -17,7 +17,10 @@ import {MatInputModule} from '@angular/material/input';
 import {MatFormFieldModule} from '@angular/material/form-field';
 
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
-import { UploadImageService } from './shared/upload-image.service';
+import { UploadImageService } from './services/upload-image.service';
+
+import { MobxAngularModule } from 'mobx-angular';
+import { GoodsStore } from './store/app.store';
 
 @NgModule({
   declarations: [
@@ -37,12 +40,13 @@ import { UploadImageService } from './shared/upload-image.service';
     NgbModule.forRoot(),
     MatInputModule,
     MatFormFieldModule,
+    MobxAngularModule
     //remove this after real API done
 /*     HttpClientInMemoryWebApiModule.forRoot(
       InMemoryDataService, { dataEncapsulation: false }
     ) */
   ],
-  providers: [SimulateDbService, UploadImageService],
+  providers: [DbService, UploadImageService, GoodsStore],
   bootstrap: [AppComponent]
 })
 export class AppModule { }
