@@ -1,6 +1,8 @@
-import { Component, OnInit, Input, Output, EventEmitter } from '@angular/core';
-import { SimulateDbService } from '../simulate-db.service';
-import { Goods } from './goods';
+import { Component, OnInit, Output, EventEmitter } from '@angular/core';
+import { Goods } from '../../models/goods';
+import { GoodsStore, Good } from 'src/app/store/app.store';
+import { Observable } from 'rxjs';
+import { fromMobx } from 'ngx-mobx';
 
 @Component({
   selector: 'app-goods',
@@ -9,8 +11,6 @@ import { Goods } from './goods';
 })
 export class GoodsComponent implements OnInit {
 
-  @Input() categoryName: string;
-  @Input() categoryId: number;
   @Output() clickCategoryId: EventEmitter<number> = new EventEmitter();
   @Output() goodFromGoodsComponent: EventEmitter<Goods> = new EventEmitter();
 
@@ -18,20 +18,18 @@ export class GoodsComponent implements OnInit {
 
   public collectionSize: number;
 
-  public goods: Goods[];
-  filteredGoods: Goods[];
-
   public itemsPerPage: number = 5;
   resourcesLoaded: boolean;
 
-  constructor(private simulateDbService: SimulateDbService) {
+  goods: Observable<Good>;
+
+  constructor(public _goodsStore: GoodsStore) {
     this.page = 1;
-    this.loadPage();
-    console.log("GoodsComponent " + this.categoryId);
+    console.log("GoodsComponent collectionSize " + this._goodsStore.counterGoods);
   }
 
   ngOnInit() {
-    this.loadPage();
+    this.goods = fromMobx(() => this._goodsStore.filtForPaginStoreGoods);
   }
 
   onPageChanged() {
@@ -39,7 +37,7 @@ export class GoodsComponent implements OnInit {
   }
 
   deleteRow(good) {
-    this.simulateDbService.deleteGood(good).subscribe();
+    this._goodsStore.removeGood(good);
     this.loadPage();
   }
   editRow(good) {
@@ -50,20 +48,18 @@ export class GoodsComponent implements OnInit {
   private loadPage() {
     this.resourcesLoaded = false;
     setTimeout(() => {
-      this.simulateDbService.getCategories(this.page, this.itemsPerPage, this.categoryId)
-        .subscribe(p => {
-          this.goods = p.rows;
-          this.collectionSize = p.totalCount;
-          this.resourcesLoaded = true;
-        });
+      this._goodsStore.page = this.page;
+      this.collectionSize = this._goodsStore.counterGoods;
+      console.log(this.page + " col size " + this.collectionSize);
+      this.resourcesLoaded = true;
     }, 500); //От этого таймаута надо избавиться! Он тут временно
-  }
+  } 
 
   newGood(){
     let emptyGood: Goods = {
       name: "",
       price: null,
-      categoryId: this.categoryId,
+      categoryId: this._goodsStore.filter,
       Qty: null,
       url: "/assets/images/placeholder-image.png",
       id: null
